extract more stuff
https://coveord.atlassian.net/browse/KIT-2076
