+++ packages/atomic/cypress/fixtures/test-fixture.ts
@@ -1,16 +1,23 @@
 import {Result} from '@coveo/headless';
 import {i18n} from 'i18next';
 import {SearchResponseSuccess} from '../../../headless/dist/definitions/api/search/search/search-response';
-import {AnalyticsTracker, AnyEventRequest} from '../utils/analyticsUtils';
 import {buildTestUrl} from '../utils/setupComponent';
 import {
   ConsoleAliases,
   getUABody,
   getUACustomData,
+  interceptSearchAndReturnError,
+  modifySearchResponses,
   RouteAlias,
+  sampleConfig,
+  SearchResponseModifier,
+  SearchResponseModifierPredicate,
   setupIntercept,
   stubConsole,
   UrlParts,
+  TestFeature,
+  configureI18n,
+  interceptAnalytics,
 } from './fixture-common';
 
 interface ResultWithFolding extends Result {
@@ -25,15 +32,6 @@ export type SearchFoldedResponseSuccess = Omit<
   results: ResultWithFolding[];
 };
 
-export type SearchResponseModifierPredicate = (
-  response: SearchResponseSuccess
-) => SearchResponseSuccess | void;
-
-export interface SearchResponseModifier {
-  predicate: SearchResponseModifierPredicate;
-  times: number;
-}
-
 export type SearchInterface = HTMLElement & {
   initialize: (opts: {
     accessToken: string;
@@ -42,8 +40,6 @@ export type SearchInterface = HTMLElement & {
   executeFirstSearch: () => void;
   i18n: i18n;
 };
-
-export type TestFeature<T> = (e: T) => void | Promise<void>;
 
 export type TagProps = Record<string, string | number>;
 
@@ -177,6 +173,12 @@ export class TestFixture {
     return this.withViewport('iphone-x');
   }
 
+  public get elementAliases() {
+    return {
+      SearchInterface: 'searchInterface',
+    };
+  }
+
   public init() {
     !this.redirected && cy.visit(buildTestUrl(this.hash));
     cy.injectAxe();
@@ -203,72 +205,31 @@ export class TestFixture {
       if (this.disabledAnalytics) {
         searchInterfaceComponent.setAttribute('analytics', 'false');
       } else {
-        AnalyticsTracker.reset();
-        cy.intercept(
-          {
-            method: 'POST',
-            url: '**/rest/ua/v15/analytics/*',
-          },
-          (request) => AnalyticsTracker.push(request.body as AnyEventRequest)
-        );
+        interceptAnalytics();
       }
 
       if (this.responseModifiers.length) {
-        interceptSearchResponse((response) => {
-          let combinedResponse = response;
-          this.responseModifiers.forEach((modifier) => {
-            if (modifier.times <= 0) {
-              return;
-            }
-            combinedResponse =
-              modifier.predicate(combinedResponse) || combinedResponse;
-            modifier.times--;
-          });
-          return combinedResponse;
-        });
+        modifySearchResponses(this.responseModifiers);
       }
 
       if (this.returnError) {
-        cy.intercept(
-          {
-            method: 'POST',
-            url: '**/rest/search/v2?*',
-          },
-          (request) =>
-            request.reply((response) =>
-              response.send(418, {
-                exception: {code: 'Something very weird just happened'},
-              })
-            )
-        ).as(TestFixture.interceptAliases.Search.substring(1));
+        interceptSearchAndReturnError();
       }
 
       if (!this.initializeInterface) {
         return;
       }
 
-      searchInterfaceComponent
-        .initialize({
-          accessToken: 'xx564559b1-0045-48e1-953c-3addd1ee4457',
-          organizationId: 'searchuisamples',
-        })
-        .then(() => {
-          this.fieldCaptions.forEach(({field, captions}) =>
-            searchInterfaceComponent.i18n.addResourceBundle(
-              'en',
-              `caption-${field}`,
-              captions
-            )
-          );
-          searchInterfaceComponent.i18n.addResourceBundle(
-            'en',
-            'translation',
-            this.translations
-          );
-          if (this.execFirstSearch) {
-            searchInterfaceComponent.executeFirstSearch();
-          }
-        });
+      searchInterfaceComponent.initialize(sampleConfig).then(() => {
+        configureI18n(
+          searchInterfaceComponent.i18n,
+          this.translations,
+          this.fieldCaptions
+        );
+        if (this.execFirstSearch) {
+          searchInterfaceComponent.executeFirstSearch();
+        }
+      });
     });
 
     if (this.execFirstSearch && this.firstIntercept) {
@@ -281,12 +242,6 @@ export class TestFixture {
     this.aliases.forEach((alias) => alias(this));
 
     return this;
-  }
-
-  public get elementAliases() {
-    return {
-      SearchInterface: 'searchInterface',
-    };
   }
 }
 
@@ -302,26 +257,3 @@ export const generateComponentHTML = (tag: string, props: TagProps = {}) => {
   }
   return e;
 };
-
-export function interceptSearchResponse(
-  modifier: SearchResponseModifierPredicate,
-  times = 9999
-) {
-  cy.intercept(
-    {
-      method: 'POST',
-      url: '**/rest/search/v2?*',
-      times,
-    },
-    (request) => {
-      request.reply((response) => {
-        let newResponse = response.body;
-        const returnedResponse = modifier(newResponse);
-        if (returnedResponse) {
-          newResponse = returnedResponse;
-        }
-        response.send(200, newResponse);
-      });
-    }
-  ).as(TestFixture.interceptAliases.Search.substring(1));
-}
